Experimental support for shorturls.
This is definitely not suitable for merging into the mainline project,
yet, but the idea is that if a page has a shorturl defined in its front
matter, that URL should be used as the webmention source instead of the
authoritative URL.

Among other things, this means that when you use the plugin with
brid.gy, it sends over the short URL to places like Twitter.

This model assumes that any inbound webmentions to the short URL will be
302-redirected to the full-length URL so that they still just work with
places like webmention.io.